lua mode tidying
