+++ app/src/mode.js
@@ -1,0 +1,10 @@
+export default class EditorMode {
+  constructor(id) {
+    this.id = id;
+  }
+
+  /* eslint-disable-next-line no-unused-vars */
+  onRender(editor) {
+    // no-op; optionally overridden in subclasses.
+  }
+}
